app/requirements.txt
azure-identity==1.13.0
azure-functions==1.15.0
azure-storage-blob==12.17.0
azure-search-documents==11.4.0b6
tenacity==8.2.2
azure-core==1.28.0
azure-storage-file-datalake==12.12.0
openai==0.27.8
pandas==2.0.3
azure-kusto-data==4.2.0